requirements.txt
pillow
rembg
trimesh
pyrender
matplotlib